Refactor into proper namespaces
