+++ src/ACLExample/ContextA/CheckingAccount/CheckingAccount.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DDDHH\ACLExample\ContextA;
+namespace DDDHH\ACLExample\ContextA\CheckingAccount;
 
 /**
  * A banking account
